+++ sdk/identity/azure-identity-broker/azure/identity/broker/_browser.py
@@ -52,8 +52,9 @@ class InteractiveBrowserBrokerCredential(_InteractiveBrowserCredential):
         may still log in with a different username.
     :paramtype cache_persistence_options: ~azure.identity.TokenCachePersistenceOptions
     :keyword int timeout: seconds to wait for the user to complete authentication. Defaults to 300 (5 minutes).
-    :keyword int parent_window_handle: If your app is a GUI app running on a modern Windows system, you are required to
-        also provide its window handle so that the sign in UI window will properly pop up on top of your window.
+    :keyword int parent_window_handle: If your app is a GUI app running on Windows 10+ or macOS, you
+        are required to also provide its window handle, so that the sign in UI window will properly pop up on top
+        of your window.
     :keyword bool use_default_broker_account: Enables automatically using the default broker account for
         authentication instead of prompting the user with an account picker. Defaults to False.
     :keyword bool enable_msa_passthrough: Determines whether Microsoft Account (MSA) passthrough is enabled. Note, this